Use safeargs for kotlin and move ActivityUtils to corecomponent.
