+++ app/build.gradle.kts
@@ -12,7 +12,7 @@ plugins {
 
     kotlin("kapt")
 
-    id("androidx.navigation.safeargs")
+    id("androidx.navigation.safeargs.kotlin")
 }
 
 android {
