--- app/src/main/res/navigation/nav_graph.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/nav_graph"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:ignore="UnusedNavigation">
-
-    <dialog android:name="com.tick.taku.notificationwatcher.view.dialog.ApplicationDialog">
-        <argument
-            android:name="model"
-            app:argType="com.tick.taku.notificationwatcher.view.dialog.ApplicationDialog.Model"/>
-    </dialog>
-
-</navigation>
